sanity check loop

+++ my-github-cleaner-base/src/main/java/me/sathish/my_github_cleaner/base/github/DBCleanupMessageConsumer.java
@@ -1,7 +1,9 @@
 package me.sathish.my_github_cleaner.base.github;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import java.time.LocalDateTime;
+
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import me.sathish.my_github_cleaner.base.eventracker.EventTrackerService;
@@ -25,6 +27,7 @@ public class DBCleanupMessageConsumer {
         this.objectMapper = objectMapper;
         this.repositoriesRepository = repositoriesRepository;
     }
+
     /*
     {"id":null,"eventId":"a07968f2-4bbf-42eb-8853-0b09b18ee5e8","eventType":"GITHUB_REPOSITORY_PROJECT","payload":"Failed to delete repository
     {\"Repo Record ID\":\"10007\",\"repositoryName\":\"gjhj\",\"deletedAt\":\"2025-09-14T08:52:25.167003\",\"deletedBy\":\"sathishjayapal\"}","createdBy":"sathishjayapal","updatedBy":"sathishjayapal","domain":10093}
@@ -48,9 +51,11 @@ public class DBCleanupMessageConsumer {
                         payload.getRepositoryName(),
                         payload.getDeletedBy(),
                         payload.getDeletedAt());
-                repositoriesRepository.deleteById(Long.parseLong(payload.getRepoRecordId()));
-                String eventPayload = createEventPayload(payload);
-                eventTrackerService.sendGitHubEventToEventstracker(eventPayload);
+                if (payload.getRepoRecordId() != null) {
+                    repositoriesRepository.deleteById(Long.parseLong(payload.getRepoRecordId()));
+                    String eventPayload = createEventPayload(payload);
+                    eventTrackerService.sendGitHubEventToEventstracker(eventPayload);
+                }
             }
         } catch (Exception e) {
             log.error("Error deserializing message: {}", message, e);
